chore: refactor integration tests
