+++ crates/zerobox/tests/integration.rs
@@ -1,0 +1,4 @@
+mod support;
+
+mod sandbox;
+mod snapshot;
