--- crates/zerobox/tests/sandbox_e2e.rs
@@ -1,1234 +1,0 @@
-//! End-to-end tests for zerobox-exec sandbox enforcement.
-//!
-//! Each test runs the real binary through the Codex sandboxing API.
-//! On macOS this exercises seatbelt; on Linux bubblewrap+seccomp.
-//!
-//! Set ZEROBOX_EXEC to test a pre-built binary (e.g. release build).
-//! Falls back to the debug binary that `cargo test` builds automatically.
-
-use std::path::PathBuf;
-use std::process::{Command, Output};
-
-fn zerobox_exec() -> PathBuf {
-    let path: PathBuf = std::env::var("ZEROBOX_EXEC")
-        .map(PathBuf::from)
-        .unwrap_or_else(|_| env!("CARGO_BIN_EXE_zerobox").into());
-    if path.is_relative() {
-        PathBuf::from(env!("CARGO_MANIFEST_DIR"))
-            .parent()
-            .and_then(|p| p.parent())
-            .expect("workspace root")
-            .join(&path)
-    } else {
-        path
-    }
-}
-
-fn run(args: &[&str]) -> Output {
-    Command::new(zerobox_exec())
-        .args(args)
-        .output()
-        .expect("failed to spawn zerobox-exec")
-}
-
-fn stdout(output: &Output) -> String {
-    String::from_utf8_lossy(&output.stdout).to_string()
-}
-
-fn stderr(output: &Output) -> String {
-    String::from_utf8_lossy(&output.stderr).to_string()
-}
-
-fn setup_tmp(name: &str) -> PathBuf {
-    let dir = PathBuf::from(format!("/tmp/zerobox-e2e-{name}"));
-    let _ = std::fs::remove_dir_all(&dir);
-    std::fs::create_dir_all(&dir).expect("setup dir");
-    dir
-}
-
-/// Helper: run curl and return the HTTP status code string.
-fn curl_status(args: &[&str], url: &str) -> (String, bool) {
-    let mut full_args: Vec<&str> = args.to_vec();
-    full_args.extend([
-        "--",
-        "curl",
-        "-s",
-        "--max-time",
-        "5",
-        "-o",
-        "/dev/null",
-        "-w",
-        "%{http_code}",
-        url,
-    ]);
-    let out = run(&full_args);
-    let code = stdout(&out).trim().to_string();
-    (code.clone(), code == "200")
-}
-
-// ═══════════════════════════════════════════════════════════════════════════
-// Default mode: full read, no write, no network
-// ═══════════════════════════════════════════════════════════════════════════
-
-#[test]
-fn default_read_succeeds() {
-    std::fs::write("/tmp/zerobox-e2e-read", "hello").expect("setup");
-    let out = run(&["--", "cat", "/tmp/zerobox-e2e-read"]);
-    assert!(out.status.success(), "stderr: {}", stderr(&out));
-    assert_eq!(stdout(&out).trim(), "hello");
-}
-
-#[test]
-fn default_write_blocked() {
-    let out = run(&[
-        "--",
-        "node",
-        "-e",
-        "try{require('fs').writeFileSync('/tmp/zerobox-e2e-wb','x');process.exit(0)}catch(e){process.exit(1)}",
-    ]);
-    assert!(!out.status.success());
-}
-
-#[test]
-fn default_network_blocked() {
-    let out = run(&[
-        "--",
-        "node",
-        "-e",
-        "fetch('https://example.com').then(()=>process.exit(0)).catch(()=>process.exit(1))",
-    ]);
-    assert!(!out.status.success());
-}
-
-// ═══════════════════════════════════════════════════════════════════════════
-// --allow-read: restrict readable paths (includes platform defaults)
-// ═══════════════════════════════════════════════════════════════════════════
-
-mod allow_read {
-    use super::*;
-
-    #[test]
-    fn allowed_path_readable() {
-        std::fs::write("/tmp/zerobox-e2e-ar", "secret").expect("setup");
-        let out = run(&["--allow-read=/tmp", "--", "cat", "/tmp/zerobox-e2e-ar"]);
-        assert!(out.status.success(), "stderr: {}", stderr(&out));
-        assert_eq!(stdout(&out).trim(), "secret");
-    }
-
-    #[test]
-    fn unallowed_path_blocked() {
-        // Create a file under $HOME which is not in platform defaults on
-        // any OS. (/tmp is in macOS seatbelt platform defaults so can't
-        // be used here.)
-        let home = std::env::var("HOME").expect("HOME not set");
-        let dir = PathBuf::from(&home).join(".zerobox-e2e-blocked");
-        let _ = std::fs::remove_dir_all(&dir);
-        std::fs::create_dir_all(&dir).expect("setup");
-        std::fs::write(dir.join("secret.txt"), "hidden").expect("setup");
-        let out = run(&[
-            "--allow-read=/var",
-            "--",
-            "cat",
-            &dir.join("secret.txt").to_string_lossy(),
-        ]);
-        let _ = std::fs::remove_dir_all(&dir);
-        assert!(
-            !out.status.success(),
-            "should be blocked, stdout: {}",
-            super::stdout(&out)
-        );
-    }
-
-    #[test]
-    fn cat_runs_with_restricted_read() {
-        std::fs::write("/tmp/zerobox-e2e-nr", "data").expect("setup");
-        let out = run(&["--allow-read=/tmp", "--", "cat", "/tmp/zerobox-e2e-nr"]);
-        assert!(out.status.success(), "stderr: {}", stderr(&out));
-        assert_eq!(stdout(&out).trim(), "data");
-    }
-
-    #[test]
-    fn multiple_read_paths() {
-        std::fs::write("/tmp/zerobox-e2e-mr", "one").expect("setup");
-        let out = run(&["--allow-read=/tmp,/var", "--", "cat", "/tmp/zerobox-e2e-mr"]);
-        assert!(out.status.success(), "stderr: {}", stderr(&out));
-        assert_eq!(stdout(&out).trim(), "one");
-    }
-}
-
-// ═══════════════════════════════════════════════════════════════════════════
-// --deny-read: carve out exceptions within allowed reads
-// ═══════════════════════════════════════════════════════════════════════════
-
-mod deny_read {
-    use super::*;
-
-    #[test]
-    fn deny_blocks_within_default_full_read() {
-        let dir = setup_tmp("dr1");
-        let secret = dir.join("private");
-        std::fs::create_dir_all(&secret).expect("setup");
-        std::fs::write(secret.join("key.txt"), "password").expect("setup");
-
-        let out = run(&[
-            &format!("--deny-read={}", secret.display()),
-            "--",
-            "node",
-            "-e",
-            &format!(
-                "try{{require('fs').readFileSync('{}/private/key.txt');console.log('ALLOWED')}}catch(e){{console.log('BLOCKED')}}",
-                dir.display()
-            ),
-        ]);
-        assert!(out.status.success(), "stderr: {}", stderr(&out));
-        assert_eq!(stdout(&out).trim(), "BLOCKED");
-    }
-
-    #[cfg(target_os = "linux")]
-    #[test]
-    fn deny_blocks_within_explicit_allow_read() {
-        let dir = setup_tmp("dr2");
-        let secret = dir.join("secret");
-        std::fs::create_dir_all(&secret).expect("setup");
-        std::fs::write(dir.join("public.txt"), "visible").expect("setup");
-        std::fs::write(secret.join("key.txt"), "hidden").expect("setup");
-
-        let out = run(&[
-            &format!("--allow-read={}", dir.display()),
-            &format!("--deny-read={}", secret.display()),
-            "--",
-            "sh",
-            "-c",
-            &format!(
-                "cat {}/public.txt 2>/dev/null && cat {}/secret/key.txt 2>/dev/null || echo DENIED",
-                dir.display(),
-                dir.display()
-            ),
-        ]);
-        assert!(out.status.success(), "stderr: {}", stderr(&out));
-        let out_str = stdout(&out);
-        assert!(
-            out_str.contains("visible"),
-            "public file should be readable, got: {out_str}"
-        );
-        assert!(
-            !out_str.contains("hidden"),
-            "secret file should be blocked, got: {out_str}"
-        );
-    }
-}
-
-// ═══════════════════════════════════════════════════════════════════════════
-// --allow-write / --deny-write
-// ═══════════════════════════════════════════════════════════════════════════
-
-#[test]
-fn allow_write_specific_path() {
-    let out = run(&[
-        "--allow-write=/tmp",
-        "--",
-        "node",
-        "-e",
-        "require('fs').writeFileSync('/tmp/zerobox-e2e-aw','written');console.log('ok')",
-    ]);
-    assert!(out.status.success(), "stderr: {}", stderr(&out));
-    assert_eq!(stdout(&out).trim(), "ok");
-    let content = std::fs::read_to_string("/tmp/zerobox-e2e-aw").expect("read back");
-    assert_eq!(content, "written");
-}
-
-#[test]
-fn allow_write_does_not_grant_other_paths() {
-    let out = run(&[
-        "--allow-write=/tmp",
-        "--",
-        "node",
-        "-e",
-        "try{require('fs').writeFileSync('/var/zerobox-e2e-aw','x');process.exit(0)}catch(e){process.exit(1)}",
-    ]);
-    assert!(!out.status.success());
-}
-
-mod deny_write {
-    use super::*;
-
-    #[test]
-    fn deny_blocks_within_allowed_dir() {
-        let dir = setup_tmp("dw1");
-        let protected = dir.join(".git");
-        std::fs::create_dir_all(&protected).expect("setup");
-
-        let out = run(&[
-            &format!("--allow-write={}", dir.display()),
-            &format!("--deny-write={}", protected.display()),
-            "--",
-            "node",
-            "-e",
-            &format!(
-                r#"
-const fs = require('fs');
-let r = [];
-try {{ fs.writeFileSync('{}/file.txt','ok'); r.push('file:ok'); }} catch(e) {{ r.push('file:blocked'); }}
-try {{ fs.writeFileSync('{}/.git/evil','x'); r.push('git:ok'); }} catch(e) {{ r.push('git:blocked'); }}
-console.log(r.join(','));
-"#,
-                dir.display(),
-                dir.display()
-            ),
-        ]);
-        assert!(out.status.success(), "stderr: {}", stderr(&out));
-        let result = stdout(&out).trim().to_string();
-        assert!(result.contains("file:ok"), "got: {result}");
-        assert!(result.contains("git:blocked"), "got: {result}");
-    }
-
-    #[test]
-    fn deny_write_with_full_write() {
-        let dir = setup_tmp("dw2");
-        std::fs::create_dir_all(&dir).expect("setup");
-
-        let out = run(&[
-            "--allow-write",
-            &format!("--deny-write={}", dir.display()),
-            "--",
-            "node",
-            "-e",
-            &format!(
-                "try{{require('fs').writeFileSync('{}/blocked.txt','x');console.log('ALLOWED')}}catch(e){{console.log('BLOCKED')}}",
-                dir.display()
-            ),
-        ]);
-        assert!(out.status.success(), "stderr: {}", stderr(&out));
-        assert_eq!(stdout(&out).trim(), "BLOCKED");
-    }
-}
-
-// ═══════════════════════════════════════════════════════════════════════════
-// --allow-net: boolean (all network)
-// ═══════════════════════════════════════════════════════════════════════════
-
-#[test]
-fn allow_net_full_permits_outbound() {
-    let (code, ok) = curl_status(&["--allow-net"], "https://example.com");
-    assert!(ok, "expected 200, got {code}");
-}
-
-// ═══════════════════════════════════════════════════════════════════════════
-// --allow-net=<domains>: domain-level filtering via Codex network proxy
-// ═══════════════════════════════════════════════════════════════════════════
-
-mod allow_net_domains {
-    use super::*;
-
-    #[test]
-    fn single_domain_allowed() {
-        let (code, ok) = curl_status(&["--allow-net=example.com"], "https://example.com");
-        assert!(ok, "expected 200, got {code}");
-    }
-
-    #[test]
-    fn unlisted_domain_blocked() {
-        let (code, ok) = curl_status(&["--allow-net=example.com"], "https://google.com");
-        assert!(!ok, "expected blocked, got {code}");
-    }
-
-    #[test]
-    fn multiple_domains_allowed() {
-        let (code, ok) = curl_status(
-            &["--allow-net=example.com,google.com"],
-            "https://example.com",
-        );
-        assert!(ok, "expected 200, got {code}");
-    }
-
-    #[test]
-    fn wildcard_subdomain_allows_subdomains() {
-        // *.example.com should allow www.example.com
-        // (we can't easily test a real subdomain that resolves, so test
-        // that the apex is NOT matched by the wildcard -- that's the
-        // important semantic)
-        let (code, ok) = curl_status(&["--allow-net=*.example.com"], "https://example.com");
-        assert!(
-            !ok,
-            "*.example.com should NOT match apex example.com, got {code}"
-        );
-    }
-
-    #[test]
-    fn allowed_domain_passes_unlisted_fails() {
-        let (code, ok) = curl_status(&["--allow-net=example.com"], "https://example.com");
-        assert!(ok, "allowed domain should pass, got {code}");
-        let (code, ok) = curl_status(&["--allow-net=example.com"], "https://google.com");
-        assert!(!ok, "unlisted domain should fail, got {code}");
-    }
-
-    #[test]
-    fn apex_and_wildcard_combined() {
-        // To allow both apex and subdomains, list both.
-        let (code, ok) = curl_status(
-            &["--allow-net=example.com,*.example.com"],
-            "https://example.com",
-        );
-        assert!(ok, "expected 200, got {code}");
-    }
-}
-
-// ═══════════════════════════════════════════════════════════════════════════
-// --deny-net=<domains>: block specific domains
-// ═══════════════════════════════════════════════════════════════════════════
-
-mod deny_net_domains {
-    use super::*;
-
-    #[test]
-    fn deny_blocks_specific_domain() {
-        let (code, ok) = curl_status(
-            &["--allow-net", "--deny-net=google.com"],
-            "https://google.com",
-        );
-        assert!(!ok, "expected blocked, got {code}");
-    }
-
-    #[test]
-    fn deny_does_not_affect_other_domains() {
-        let (code, ok) = curl_status(
-            &["--allow-net", "--deny-net=google.com"],
-            "https://example.com",
-        );
-        assert!(ok, "expected 200, got {code}");
-    }
-
-    #[test]
-    fn deny_overrides_allow() {
-        // Allow example.com but also deny it. Deny wins.
-        let (code, ok) = curl_status(
-            &["--allow-net=example.com", "--deny-net=example.com"],
-            "https://example.com",
-        );
-        assert!(!ok, "deny should override allow, got {code}");
-    }
-
-    #[test]
-    fn deny_wildcard_blocks_subdomains() {
-        // Deny *.google.com, allow everything else.
-        let (code, ok) = curl_status(
-            &["--allow-net", "--deny-net=*.google.com"],
-            "https://example.com",
-        );
-        assert!(ok, "example.com should still work, got {code}");
-    }
-
-    #[test]
-    fn deny_multiple_domains() {
-        // Deny both google.com and example.com.
-        let (code, ok) = curl_status(
-            &["--allow-net", "--deny-net=google.com,example.com"],
-            "https://example.com",
-        );
-        assert!(!ok, "example.com should be denied, got {code}");
-    }
-}
-
-// ═══════════════════════════════════════════════════════════════════════════
-// --allow-all / --no-sandbox
-// ═══════════════════════════════════════════════════════════════════════════
-
-#[test]
-fn allow_all_permits_everything() {
-    let out = run(&[
-        "--allow-all",
-        "--",
-        "node",
-        "-e",
-        "require('fs').writeFileSync('/tmp/zerobox-e2e-aa','x');console.log('ok')",
-    ]);
-    assert!(out.status.success(), "stderr: {}", stderr(&out));
-    assert_eq!(stdout(&out).trim(), "ok");
-}
-
-#[test]
-fn no_sandbox_permits_everything() {
-    let out = run(&[
-        "--no-sandbox",
-        "--",
-        "node",
-        "-e",
-        "require('fs').writeFileSync('/tmp/zerobox-e2e-ns','x');console.log('ok')",
-    ]);
-    assert!(out.status.success(), "stderr: {}", stderr(&out));
-    assert_eq!(stdout(&out).trim(), "ok");
-}
-
-// ═══════════════════════════════════════════════════════════════════════════
-// Combined flags
-// ═══════════════════════════════════════════════════════════════════════════
-
-#[test]
-fn allow_read_and_write_combined() {
-    std::fs::write("/tmp/zerobox-e2e-rw-in", "input").expect("setup");
-    let out = run(&[
-        "--allow-read=/tmp",
-        "--allow-write=/tmp",
-        "--",
-        "sh",
-        "-c",
-        "cat /tmp/zerobox-e2e-rw-in > /tmp/zerobox-e2e-rw-out && echo ok",
-    ]);
-    assert!(out.status.success(), "stderr: {}", stderr(&out));
-    assert_eq!(stdout(&out).trim(), "ok");
-    let content = std::fs::read_to_string("/tmp/zerobox-e2e-rw-out").expect("read back");
-    assert_eq!(content.trim(), "input");
-}
-
-#[test]
-fn allow_read_and_net_combined() {
-    // /run is needed on systemd-based Linux because /etc/resolv.conf
-    // symlinks to /run/systemd/resolve/stub-resolv.conf for DNS.
-    let (code, ok) = curl_status(
-        &[
-            "--allow-read=/tmp,/run",
-            "--allow-write=/tmp",
-            "--allow-net",
-        ],
-        "https://example.com",
-    );
-    assert!(ok, "expected 200, got {code}");
-}
-
-#[test]
-fn deny_read_and_deny_write_combined() {
-    let dir = setup_tmp("combo");
-    let secret = dir.join("secret");
-    std::fs::create_dir_all(&secret).expect("setup");
-    std::fs::write(dir.join("public"), "hello").expect("setup");
-
-    let out = run(&[
-        &format!("--allow-write={}", dir.display()),
-        &format!("--deny-write={}", secret.display()),
-        "--",
-        "node",
-        "-e",
-        &format!(
-            r#"
-const fs = require('fs');
-let r = [];
-try {{ fs.writeFileSync('{}/new.txt','x'); r.push('write-pub:ok'); }} catch(e) {{ r.push('write-pub:blocked'); }}
-try {{ fs.writeFileSync('{}/secret/evil','x'); r.push('write-sec:ok'); }} catch(e) {{ r.push('write-sec:blocked'); }}
-console.log(r.join(','));
-"#,
-            dir.display(),
-            dir.display()
-        ),
-    ]);
-    assert!(out.status.success(), "stderr: {}", stderr(&out));
-    let result = stdout(&out).trim().to_string();
-    assert!(result.contains("write-pub:ok"), "got: {result}");
-    assert!(result.contains("write-sec:blocked"), "got: {result}");
-}
-
-#[test]
-fn allow_net_domain_with_write_restriction() {
-    let dir = setup_tmp("net-write");
-    let (code, ok) = curl_status(
-        &[
-            "--allow-net=example.com",
-            &format!("--allow-write={}", dir.display()),
-        ],
-        "https://example.com",
-    );
-    assert!(ok, "expected 200, got {code}");
-}
-
-// ═══════════════════════════════════════════════════════════════════════════
-// Exit code propagation
-// ═══════════════════════════════════════════════════════════════════════════
-
-#[test]
-fn exit_code_zero_propagated() {
-    let out = run(&["--", "node", "-e", "process.exit(0)"]);
-    assert!(out.status.success());
-}
-
-#[test]
-fn exit_code_nonzero_propagated() {
-    let out = run(&["--", "node", "-e", "process.exit(42)"]);
-    assert_eq!(out.status.code(), Some(42));
-}
-
-// ═══════════════════════════════════════════════════════════════════════════
-// Edge cases
-// ═══════════════════════════════════════════════════════════════════════════
-
-#[test]
-fn relative_write_path_resolved() {
-    let out = run(&[
-        "--allow-write=/tmp",
-        "-C",
-        "/tmp",
-        "--",
-        "node",
-        "-e",
-        "require('fs').writeFileSync('/tmp/zerobox-e2e-rel','ok');console.log('ok')",
-    ]);
-    assert!(out.status.success(), "stderr: {}", stderr(&out));
-}
-
-#[test]
-fn nonexistent_command_fails() {
-    let out = run(&["--", "this-command-does-not-exist-zerobox"]);
-    assert!(!out.status.success());
-}
-
-// ═══════════════════════════════════════════════════════════════════════════
-// Environment variables: --env, --allow-env, --deny-env
-// ═══════════════════════════════════════════════════════════════════════════
-
-mod env_vars {
-    use super::*;
-
-    // ── Default: clean env (only essentials) ──
-
-    #[test]
-    fn default_env_has_path() {
-        let out = run(&["--allow-all", "--", "sh", "-c", "echo $PATH"]);
-        assert!(out.status.success());
-        assert!(!stdout(&out).trim().is_empty(), "PATH should be set");
-    }
-
-    #[test]
-    fn default_env_has_home() {
-        let out = run(&["--allow-all", "--", "sh", "-c", "echo $HOME"]);
-        assert!(out.status.success());
-        assert!(!stdout(&out).trim().is_empty(), "HOME should be set");
-    }
-
-    #[test]
-    fn default_env_excludes_non_essential_vars() {
-        // Set a custom env var in parent, verify it does NOT reach the child.
-        let out = Command::new(zerobox_exec())
-            .args(["--allow-all", "--", "sh", "-c", "echo $ZEROBOX_TEST_VAR"])
-            .env("ZEROBOX_TEST_VAR", "should_not_appear")
-            .output()
-            .expect("spawn");
-        assert!(out.status.success());
-        assert_eq!(
-            stdout(&out).trim(),
-            "",
-            "custom var should not be inherited"
-        );
-    }
-
-    #[test]
-    fn default_env_is_minimal() {
-        let out = run(&["--allow-all", "--", "env"]);
-        assert!(out.status.success());
-        let count = stdout(&out).lines().count();
-        assert!(count <= 10, "expected minimal env, got {count} vars");
-    }
-
-    // ── --env KEY=VALUE ──
-
-    #[test]
-    fn env_sets_explicit_var() {
-        let out = run(&[
-            "--allow-all",
-            "--env",
-            "FOO=bar",
-            "--",
-            "sh",
-            "-c",
-            "echo $FOO",
-        ]);
-        assert!(out.status.success());
-        assert_eq!(stdout(&out).trim(), "bar");
-    }
-
-    #[test]
-    fn env_multiple_vars() {
-        let out = run(&[
-            "--allow-all",
-            "--env",
-            "A=1",
-            "--env",
-            "B=2",
-            "--",
-            "sh",
-            "-c",
-            "echo $A $B",
-        ]);
-        assert!(out.status.success());
-        assert_eq!(stdout(&out).trim(), "1 2");
-    }
-
-    #[test]
-    fn env_value_with_equals() {
-        let out = run(&[
-            "--allow-all",
-            "--env",
-            "DATA=key=value=extra",
-            "--",
-            "sh",
-            "-c",
-            "echo $DATA",
-        ]);
-        assert!(out.status.success());
-        assert_eq!(stdout(&out).trim(), "key=value=extra");
-    }
-
-    #[test]
-    fn env_empty_value() {
-        let out = run(&[
-            "--allow-all",
-            "--env",
-            "EMPTY=",
-            "--",
-            "sh",
-            "-c",
-            "echo \"x${EMPTY}x\"",
-        ]);
-        assert!(out.status.success());
-        assert_eq!(stdout(&out).trim(), "xx");
-    }
-
-    #[test]
-    fn env_overrides_inherited() {
-        let out = run(&[
-            "--allow-all",
-            "--allow-env",
-            "--env",
-            "HOME=/custom",
-            "--",
-            "sh",
-            "-c",
-            "echo $HOME",
-        ]);
-        assert!(out.status.success());
-        assert_eq!(stdout(&out).trim(), "/custom");
-    }
-
-    // ── --allow-env (inherit all) ──
-
-    #[test]
-    fn allow_env_bare_inherits_all() {
-        let out = Command::new(zerobox_exec())
-            .args([
-                "--allow-all",
-                "--allow-env",
-                "--",
-                "sh",
-                "-c",
-                "echo $ZEROBOX_TEST_ALL",
-            ])
-            .env("ZEROBOX_TEST_ALL", "visible")
-            .output()
-            .expect("spawn");
-        assert!(out.status.success());
-        assert_eq!(stdout(&out).trim(), "visible");
-    }
-
-    #[test]
-    fn allow_env_bare_has_many_vars() {
-        let out = run(&["--allow-all", "--allow-env", "--", "env"]);
-        assert!(out.status.success());
-        let count = stdout(&out).lines().count();
-        assert!(
-            count > 10,
-            "expected many vars with --allow-env, got {count}"
-        );
-    }
-
-    // ── --allow-env=KEY1,KEY2 ──
-
-    #[test]
-    fn allow_env_specific_keys() {
-        let out = run(&["--allow-all", "--allow-env=PATH", "--", "env"]);
-        assert!(out.status.success());
-        let lines = stdout(&out);
-        assert!(lines.contains("PATH="), "PATH should be present");
-        assert!(!lines.contains("HOME="), "HOME should not be present");
-    }
-
-    #[test]
-    fn allow_env_specific_with_env_override() {
-        // --allow-env=PATH filters to only PATH, but --env adds FOO.
-        let out = run(&[
-            "--allow-all",
-            "--allow-env=PATH",
-            "--env",
-            "FOO=added",
-            "--",
-            "sh",
-            "-c",
-            "echo FOO=$FOO HOME=$HOME",
-        ]);
-        assert!(out.status.success());
-        let s = stdout(&out).trim().to_string();
-        assert!(
-            s.contains("FOO=added"),
-            "explicit --env should survive, got: {s}"
-        );
-        assert!(s.contains("HOME="), "HOME should be empty");
-        assert!(!s.contains("HOME=/"), "HOME should not have a value");
-    }
-
-    // ── --deny-env ──
-
-    #[test]
-    fn deny_env_removes_var() {
-        let out = run(&[
-            "--allow-all",
-            "--allow-env",
-            "--deny-env=HOME",
-            "--",
-            "sh",
-            "-c",
-            "echo \"HOME=$HOME\"",
-        ]);
-        assert!(out.status.success());
-        assert_eq!(stdout(&out).trim(), "HOME=");
-    }
-
-    #[test]
-    fn deny_env_takes_precedence_over_allow() {
-        let out = run(&[
-            "--allow-all",
-            "--allow-env=PATH,HOME",
-            "--deny-env=HOME",
-            "--",
-            "env",
-        ]);
-        assert!(out.status.success());
-        let lines = stdout(&out);
-        assert!(lines.contains("PATH="), "PATH should survive");
-        assert!(!lines.contains("HOME="), "HOME should be denied");
-    }
-
-    #[test]
-    fn deny_env_does_not_block_explicit_env() {
-        // --deny-env=FOO + --env FOO=override → FOO survives.
-        let out = run(&[
-            "--allow-all",
-            "--allow-env",
-            "--deny-env=FOO",
-            "--env",
-            "FOO=override",
-            "--",
-            "sh",
-            "-c",
-            "echo $FOO",
-        ]);
-        assert!(out.status.success());
-        assert_eq!(stdout(&out).trim(), "override");
-    }
-
-    #[test]
-    fn deny_env_multiple_keys() {
-        let out = run(&[
-            "--allow-all",
-            "--allow-env",
-            "--deny-env=HOME,USER",
-            "--",
-            "sh",
-            "-c",
-            "echo \"HOME=$HOME USER=$USER\"",
-        ]);
-        assert!(out.status.success());
-        let s = stdout(&out).trim().to_string();
-        assert_eq!(s, "HOME= USER=");
-    }
-
-    // ── Error cases ──
-
-    #[test]
-    fn env_without_equals_fails() {
-        let out = run(&["--allow-all", "--env", "BADFORMAT", "--", "echo", "hi"]);
-        assert!(!out.status.success());
-        assert!(
-            stderr(&out).contains("KEY=VALUE"),
-            "should hint at format, got: {}",
-            stderr(&out)
-        );
-    }
-}
-
-// ═══════════════════════════════════════════════════════════════════════════
-// Secrets: --secret, --secret-host
-// ═══════════════════════════════════════════════════════════════════════════
-
-mod secrets {
-    use super::*;
-
-    #[test]
-    fn secret_sets_placeholder_in_env() {
-        let out = run(&[
-            "--allow-all",
-            "--secret",
-            "MY_TOKEN=real-value",
-            "--secret-host",
-            "MY_TOKEN=httpbin.org",
-            "--",
-            "sh",
-            "-c",
-            "echo $MY_TOKEN",
-        ]);
-        assert!(out.status.success(), "stderr: {}", stderr(&out));
-        let val = stdout(&out).trim().to_string();
-        assert!(
-            val.starts_with("ZEROBOX_SECRET_"),
-            "expected placeholder, got: {val}"
-        );
-        assert_ne!(val, "real-value", "real value must not be visible");
-    }
-
-    #[test]
-    fn secret_without_host_works() {
-        let out = run(&[
-            "--allow-all",
-            "--secret",
-            "TOKEN=abc",
-            "--",
-            "sh",
-            "-c",
-            "echo $TOKEN",
-        ]);
-        assert!(out.status.success(), "stderr: {}", stderr(&out));
-        let val = stdout(&out).trim().to_string();
-        assert!(val.starts_with("ZEROBOX_SECRET_"));
-    }
-
-    #[test]
-    fn secret_bad_format_fails() {
-        let out = run(&["--no-sandbox", "--secret", "NOEQUALS", "--", "echo", "hi"]);
-        assert!(!out.status.success());
-        assert!(
-            stderr(&out).contains("KEY=VALUE"),
-            "should hint at format, got: {}",
-            stderr(&out)
-        );
-    }
-
-    #[test]
-    fn secret_duplicate_key_fails() {
-        let out = run(&[
-            "--allow-all",
-            "--secret",
-            "K=v1",
-            "--secret",
-            "K=v2",
-            "--",
-            "echo",
-            "hi",
-        ]);
-        assert!(!out.status.success());
-        assert!(
-            stderr(&out).contains("duplicate"),
-            "should report duplicate, got: {}",
-            stderr(&out)
-        );
-    }
-
-    #[test]
-    fn secret_host_unknown_key_fails() {
-        let out = run(&[
-            "--allow-all",
-            "--secret",
-            "A=val",
-            "--secret-host",
-            "B=host.com",
-            "--",
-            "echo",
-            "hi",
-        ]);
-        assert!(!out.status.success());
-        assert!(
-            stderr(&out).contains("unknown"),
-            "should report unknown key, got: {}",
-            stderr(&out)
-        );
-    }
-
-    #[test]
-    fn secret_combined_with_env() {
-        // --env and --secret should not interfere with each other.
-        let out = run(&[
-            "--allow-all",
-            "--secret",
-            "TOKEN=secret",
-            "--secret-host",
-            "TOKEN=httpbin.org",
-            "--env",
-            "FOO=bar",
-            "--",
-            "sh",
-            "-c",
-            "echo FOO=$FOO TOKEN=$TOKEN",
-        ]);
-        assert!(out.status.success(), "stderr: {}", stderr(&out));
-        let s = stdout(&out).trim().to_string();
-        assert!(s.contains("FOO=bar"), "env should work, got: {s}");
-        assert!(
-            s.contains("TOKEN=ZEROBOX_SECRET_"),
-            "secret should be placeholder, got: {s}"
-        );
-    }
-
-    #[test]
-    fn secret_host_only_enables_that_host() {
-        // --secret with --secret-host should ONLY allow the specified host,
-        // not all network traffic. Other hosts must be blocked.
-        let (code, ok) = curl_status(
-            &[
-                "--secret",
-                "TOKEN=val",
-                "--secret-host",
-                "TOKEN=httpbin.org",
-            ],
-            "https://example.com",
-        );
-        assert!(!ok, "example.com should be blocked, got {code}");
-    }
-}
-
-// ═══════════════════════════════════════════════════════════════════════════
-// --strict flag
-// ═══════════════════════════════════════════════════════════════════════════
-
-mod strict_flag {
-    use super::*;
-
-    #[test]
-    fn strict_works_when_namespaces_available() {
-        // On CI (real Linux or macOS), namespaces are available.
-        // --strict should succeed, not error.
-        let out = run(&["--strict-sandbox", "--", "echo", "hello"]);
-        assert!(
-            out.status.success(),
-            "strict should work when namespaces are available, stderr: {}",
-            stderr(&out)
-        );
-        assert_eq!(stdout(&out).trim(), "hello");
-    }
-
-    #[test]
-    fn strict_with_allow_write() {
-        let out = run(&[
-            "--strict-sandbox",
-            "--allow-write=/tmp",
-            "--",
-            "sh",
-            "-c",
-            "echo ok > /tmp/zerobox-strict-test && cat /tmp/zerobox-strict-test",
-        ]);
-        assert!(out.status.success(), "stderr: {}", stderr(&out));
-        assert_eq!(stdout(&out).trim(), "ok");
-    }
-}
-
-mod snapshot {
-    use super::*;
-    use std::fs;
-
-    fn temp_dir() -> tempfile::TempDir {
-        tempfile::tempdir().expect("create temp dir")
-    }
-
-    fn run_with_home(home: &std::path::Path, args: &[&str]) -> Output {
-        Command::new(zerobox_exec())
-            .args(args)
-            .env("ZEROBOX_HOME", home)
-            .output()
-            .expect("failed to spawn zerobox")
-    }
-
-    #[test]
-    fn snapshot_records_changes() {
-        let dir = temp_dir();
-        let home = temp_dir();
-        fs::write(dir.path().join("file.txt"), "original").unwrap();
-
-        let out = run_with_home(
-            home.path(),
-            &[
-                "--snapshot",
-                &format!("--snapshot-path={}", dir.path().display()),
-                &format!("--allow-write={}", dir.path().display()),
-                "--",
-                "sh",
-                "-c",
-                &format!("echo modified > {}/file.txt", dir.path().display()),
-            ],
-        );
-        assert!(out.status.success(), "stderr: {}", stderr(&out));
-        let err = stderr(&out);
-        assert!(
-            err.contains("1 modified"),
-            "expected change summary, got: {err}"
-        );
-        assert_eq!(
-            fs::read_to_string(dir.path().join("file.txt"))
-                .unwrap()
-                .trim(),
-            "modified"
-        );
-    }
-
-    #[test]
-    fn restore_undoes_changes() {
-        let dir = temp_dir();
-        let home = temp_dir();
-        fs::write(dir.path().join("file.txt"), "original").unwrap();
-
-        let out = run_with_home(
-            home.path(),
-            &[
-                "--restore",
-                &format!("--snapshot-path={}", dir.path().display()),
-                &format!("--allow-write={}", dir.path().display()),
-                "--",
-                "sh",
-                "-c",
-                &format!(
-                    "echo modified > {}/file.txt; echo new > {}/new.txt",
-                    dir.path().display(),
-                    dir.path().display()
-                ),
-            ],
-        );
-        assert!(out.status.success(), "stderr: {}", stderr(&out));
-        let err = stderr(&out);
-        assert!(
-            err.contains("restored"),
-            "expected restore message, got: {err}"
-        );
-        assert_eq!(
-            fs::read_to_string(dir.path().join("file.txt"))
-                .unwrap()
-                .trim(),
-            "original"
-        );
-        assert!(!dir.path().join("new.txt").exists());
-    }
-
-    #[test]
-    fn snapshot_no_changes_reports_none() {
-        let dir = temp_dir();
-        let home = temp_dir();
-        fs::write(dir.path().join("file.txt"), "unchanged").unwrap();
-
-        let out = run_with_home(
-            home.path(),
-            &[
-                "--snapshot",
-                &format!("--snapshot-path={}", dir.path().display()),
-                "--",
-                "true",
-            ],
-        );
-        assert!(out.status.success(), "stderr: {}", stderr(&out));
-        let err = stderr(&out);
-        assert!(err.contains("no changes detected"), "got: {err}");
-    }
-
-    #[test]
-    fn snapshot_session_discoverable_via_list() {
-        let dir = temp_dir();
-        let home = temp_dir();
-        fs::write(dir.path().join("file.txt"), "data").unwrap();
-
-        let out = run_with_home(
-            home.path(),
-            &[
-                "--snapshot",
-                &format!("--snapshot-path={}", dir.path().display()),
-                &format!("--allow-write={}", dir.path().display()),
-                "--",
-                "sh",
-                "-c",
-                &format!("echo changed > {}/file.txt", dir.path().display()),
-            ],
-        );
-        assert!(out.status.success(), "stderr: {}", stderr(&out));
-
-        let list = run_with_home(home.path(), &["snapshot", "list"]);
-        assert!(list.status.success(), "stderr: {}", stderr(&list));
-        let list_out = stdout(&list);
-        assert!(
-            list_out.contains("sh"),
-            "expected session in list, got: {list_out}"
-        );
-    }
-
-    #[test]
-    fn snapshot_subcommand_restore_works() {
-        let dir = temp_dir();
-        let home = temp_dir();
-        fs::write(dir.path().join("file.txt"), "original").unwrap();
-
-        let out = run_with_home(
-            home.path(),
-            &[
-                "--snapshot",
-                &format!("--snapshot-path={}", dir.path().display()),
-                &format!("--allow-write={}", dir.path().display()),
-                "--",
-                "sh",
-                "-c",
-                &format!("echo changed > {}/file.txt", dir.path().display()),
-            ],
-        );
-        assert!(out.status.success(), "stderr: {}", stderr(&out));
-        assert_eq!(
-            fs::read_to_string(dir.path().join("file.txt"))
-                .unwrap()
-                .trim(),
-            "changed"
-        );
-
-        let list = run_with_home(home.path(), &["snapshot", "list"]);
-        let list_out = stdout(&list);
-        let session_id = list_out
-            .lines()
-            .nth(1) // skip header
-            .and_then(|l| l.split_whitespace().next())
-            .expect("no session in list");
-
-        let restore = run_with_home(home.path(), &["snapshot", "restore", session_id]);
-        assert!(restore.status.success(), "stderr: {}", stderr(&restore));
-        assert_eq!(
-            fs::read_to_string(dir.path().join("file.txt"))
-                .unwrap()
-                .trim(),
-            "original"
-        );
-    }
-
-    #[test]
-    fn restore_with_excluded_dir_preserves_it() {
-        let dir = temp_dir();
-        let home = temp_dir();
-        fs::write(dir.path().join("file.txt"), "original").unwrap();
-        fs::create_dir_all(dir.path().join(".git")).unwrap();
-        fs::write(dir.path().join(".git/HEAD"), "ref: refs/heads/main").unwrap();
-
-        let out = run_with_home(
-            home.path(),
-            &[
-                "--restore",
-                &format!("--snapshot-path={}", dir.path().display()),
-                &format!("--allow-write={}", dir.path().display()),
-                "--",
-                "sh",
-                "-c",
-                &format!("echo modified > {}/file.txt", dir.path().display()),
-            ],
-        );
-        assert!(out.status.success(), "stderr: {}", stderr(&out));
-        assert_eq!(
-            fs::read_to_string(dir.path().join("file.txt"))
-                .unwrap()
-                .trim(),
-            "original"
-        );
-        assert!(dir.path().join(".git/HEAD").exists());
-    }
-}
