+++ crates/zerobox/tests/snapshot/mod.rs
@@ -1,0 +1,1 @@
+mod tests;
